// src/types/product.type.ts
export interface ProductDimensions {
	width: number;
	height: number;
	depth: number;
}

export interface ProductReview {
	rating: number;
	comment: string;
	date: string;
	reviewerName: string;
	reviewerEmail: string;
}

export interface ProductMeta {
	createdAt: string;
	updatedAt: string;
	barcode: string;
	qrCode: string;
}

export interface Product {
	id: number;
	title: string;
	description: string;
	category: string;
	price: number;
	discountPercentage: number;
	rating: number;
	stock: number;
	tags: string[];
	brand: string;
	sku: string;
	weight: number;
	dimensions: ProductDimensions;
	warrantyInformation: string;
	shippingInformation: string;
	availabilityStatus: string;
	reviews: ProductReview[];
	returnPolicy: string;
	minimumOrderQuantity: number;
	meta: ProductMeta;
	thumbnail: string;
	images: string[];
}

export interface ProductsResponse {
	products: Product[];
	total: number;
	skip: number;
	limit: number;
}

export interface GetProductParams {
    limit?: number;
    skip?: number;
    search?: string;
    select?: string;
    sortBy?: string;
    order?: "asc" | "desc";
}